+++ example-dogs/src/main/java/com/zhuinden/simplestackcomposedogexample/app/MainActivity.kt
@@ -1,0 +1,52 @@
+package com.zhuinden.simplestackcomposedogexample.app
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.Modifier
+import com.zhuinden.simplestack.AsyncStateChanger
+import com.zhuinden.simplestack.History
+import com.zhuinden.simplestack.navigator.Navigator
+import com.zhuinden.simplestackcomposedogexample.features.doglist.DogListKey
+import com.zhuinden.simplestackcomposeintegration.core.AnimatingComposeStateChanger
+import com.zhuinden.simplestackcomposeintegration.core.BackstackProvider
+import com.zhuinden.simplestackextensions.navigatorktx.androidContentFrame
+import com.zhuinden.simplestackextensions.services.DefaultServiceProvider
+
+class MainActivity : AppCompatActivity() {
+    private val composeStateChanger = AnimatingComposeStateChanger()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val app = application as CustomApplication
+
+        val backstack = Navigator.configure()
+            .setGlobalServices(app.globalServices)
+            .setScopedServices(DefaultServiceProvider())
+            .setStateChanger(AsyncStateChanger(composeStateChanger))
+            .install(this, androidContentFrame, History.of(DogListKey()))
+
+        setContent {
+            BackstackProvider(backstack) {
+                MaterialTheme {
+                    Surface(color = MaterialTheme.colors.background) {
+                        Box(Modifier.fillMaxSize()) {
+                            composeStateChanger.RenderScreen()
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    override fun onBackPressed() {
+        if (!Navigator.onBackPressed(this)) {
+            super.onBackPressed()
+        }
+    }
+}