+++ example-dogs/src/main/java/com/zhuinden/simplestackcomposedogexample/core/models/ModelExtensions.kt
@@ -1,0 +1,8 @@
+package com.zhuinden.simplestackcomposedogexample.core.models
+
+import com.zhuinden.simplestackcomposedogexample.data.models.Dog
+
+fun Dog.contentDescription(): String = run {
+    val dog = this
+    "${dog.name}, ${dog.determinedSex.name}, ${dog.breed}, ${dog.age} years old"
+}