+++ example-dogs/src/main/java/com/zhuinden/simplestackcomposedogexample/core/navigation/ComposeKey.kt
@@ -1,0 +1,13 @@
+package com.zhuinden.simplestackcomposedogexample.core.navigation
+
+import android.os.Parcelable
+import com.zhuinden.simplestack.ServiceBinder
+import com.zhuinden.simplestackcomposeintegration.core.DefaultComposeKey
+import com.zhuinden.simplestackextensions.services.DefaultServiceProvider
+
+abstract class ComposeKey : DefaultComposeKey(), Parcelable, DefaultServiceProvider.HasServices {
+    override fun getScopeTag(): String = javaClass.name
+
+    override fun bindServices(serviceBinder: ServiceBinder) {
+    }
+}