+++ example-dogs/src/main/java/com/zhuinden/simplestackcomposedogexample/utils/OptionalWrapper.kt
@@ -1,0 +1,7 @@
+package com.zhuinden.simplestackcomposedogexample.utils
+
+class OptionalWrapper<T>(val value: T?) {
+    companion object {
+        fun <T> absent(): OptionalWrapper<T> = OptionalWrapper(null)
+    }
+}