+++ example/src/main/java/com/zhuinden/firstcomposeapp/Utils.kt
@@ -1,8 +1,0 @@
-package com.zhuinden.firstcomposeapp
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.showToast(message: String, duration: Int = Toast.LENGTH_SHORT) {
-    Toast.makeText(this, message, duration).show()
-}